Refactor tiered store map integration code
