+++ hazelcast/src/main/java/com/hazelcast/map/impl/operation/GetOperation.java
@@ -58,8 +58,11 @@ public final class GetOperation extends ReadonlyKeyBasedMapOperation implements 
 
     @Override
     protected void afterRunInternal() {
-        mapServiceContext.interceptAfterGet(mapContainer.getInterceptorRegistry(), result);
-        super.afterRunInternal();
+        try {
+            mapServiceContext.interceptAfterGet(mapContainer.getInterceptorRegistry(), result);
+        } finally {
+            super.afterRunInternal();
+        }
     }
 
     @Override
