+++ hazelcast/src/main/java/com/hazelcast/spi/impl/operationexecutor/OperationRunner.java
@@ -170,7 +170,6 @@ public abstract class OperationRunner {
             operation.afterRun();
         } finally {
             operation.popThreadContext();
-            operation.afterRunFinal();
         }
     }
 }
